Trying  To get the anytimeEgoAloModel working.
The premise is that the alo can be trained by one of 3 different ego learners.
Each ego learner will have a different size of state space from the others. A Fast learner that is inaccurate to a slow learner that is precise.

Currently there appears to be a bug as the agent keeps performing the same action repeatedly. The state sizes are growing from this either.